feat: model folder reorg (#195)
* reorg of models into separate files

* add additional error handling, conversion for felt

* clean type conversion
